backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
newspaper3k
pydantic
python-multipart
beautifulsoup4==4.12.2
lxml
nltk
requests
aiohttp
google-generativeai>=0.3.2